Correct non-package installation path
Packages should install to /usr and non-packages to /usr/local

By default, cmake will set the install prefix to /usr/local (yay!) but
we can override this at configuration time using the -D switch.

cmake -DCMAKE_PREFIX_PATH:PATH=<path>

When packaging we need to ensure install points to /usr, which is the
correct place for packaged files to be installed. When installing via
cmake files get installed to /usr/local, as expected

Change-Id: I08d3d6bb21d3040a79caf552e85e14c869079d79