Declare size variable
For clang this code fails to compile, and it's not hard to see why. Frankly it is amazing it compiled with GCC and CL! What I think happened is the address of the size function from the class was used. You may want to look into that, perhaps this code is never called?